refactor: replace interfaces from component to a separate file
