+++ src/pages/styles/Form.module.scss
@@ -2,9 +2,10 @@
   width: 600px;
   margin: 50px auto;
   padding: 20px 30px;
-  box-shadow: 0px 5px 12px 3px rgba(0,0,0,0.32);
+  box-shadow: 0px 0px 1px 3px rgb(150, 186, 186) inset,
+  0px 5px 5px 5px rgba(0, 0, 0, 0.32);
   border-radius: 10px;
-  background-color: rgb(190, 235, 235)  ; 
+  background-color: rgb(190, 235, 235);
 }
 
 .userCardsWrapper {
